Rename rsassa_pss.h to internal.h.
